full

+++ src/main/java/com/musobek/clickapp/repo/TaskHistoryRepository.java
@@ -3,5 +3,5 @@ package com.musobek.clickapp.repo;
 import com.musobek.clickapp.entity.TaskHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskHistoryRepository extends JpaRepository<TaskHistory, Integer> {
+public interface TaskHistoryRepository extends JpaRepository<TaskHistory, Long> {
 }
